+++ app/build.gradle
@@ -50,6 +50,8 @@ dependencies {
 
     implementation 'androidx.core:core-ktx:1.9.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
+    //by viewModel
+    implementation "androidx.fragment:fragment-ktx:1.5.6"
     implementation 'com.google.android.material:material:1.8.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
