+++ app/src/main/java/com/example/android2023/data/Constants.kt
@@ -3,6 +3,9 @@ package com.example.android2023.data
 object Constants {
 
     const val BASE_URL = "https://api.openweathermap.org/data/2.5/"
-    const val API_KEY = "29c75b1fd8b9506a7fbf478c3b766c77"
+    const val API_KEY = "d9adcf9ec0f5c5d12ce301f01b6e73f4"
 
+    //костыль
+    const val KAZAN_LATITUDE = 55.7887
+    const val KAZAN_LONGITUDE = 49.1221
 }
