--- app/src/main/java/com/example/android2023/data/response/weatherResponse/Clouds.kt
@@ -1,9 +1,0 @@
-package com.example.android2023.data.response.weatherResponse
-
-
-import com.google.gson.annotations.SerializedName
-
-data class Clouds(
-    @SerializedName("all")
-    val all: Int
-)
