--- app/src/main/java/com/example/android2023/data/response/weatherResponse/Coord.kt
@@ -1,11 +1,0 @@
-package com.example.android2023.data.response.weatherResponse
-
-
-import com.google.gson.annotations.SerializedName
-
-data class Coord(
-    @SerializedName("lat")
-    val lat: Double,
-    @SerializedName("lon")
-    val lon: Double
-)
