--- app/src/main/java/com/example/android2023/data/response/weatherResponse/Main.kt
@@ -1,23 +1,0 @@
-package com.example.android2023.data.response.weatherResponse
-
-
-import com.google.gson.annotations.SerializedName
-
-data class Main(
-    @SerializedName("feels_like")
-    val feelsLike: Double,
-    @SerializedName("grnd_level")
-    val grndLevel: Int,
-    @SerializedName("humidity")
-    val humidity: Int,
-    @SerializedName("pressure")
-    val pressure: Int,
-    @SerializedName("sea_level")
-    val seaLevel: Int,
-    @SerializedName("temp")
-    val temp: Double,
-    @SerializedName("temp_max")
-    val tempMax: Double,
-    @SerializedName("temp_min")
-    val tempMin: Double
-)
