--- app/src/main/java/com/example/android2023/data/response/weatherResponse/Sys.kt
@@ -1,17 +1,0 @@
-package com.example.android2023.data.response.weatherResponse
-
-
-import com.google.gson.annotations.SerializedName
-
-data class Sys(
-    @SerializedName("country")
-    val country: String,
-    @SerializedName("id")
-    val id: Int,
-    @SerializedName("sunrise")
-    val sunrise: Int,
-    @SerializedName("sunset")
-    val sunset: Int,
-    @SerializedName("type")
-    val type: Int
-)
