--- app/src/main/java/com/example/android2023/data/response/weatherResponse/Weather.kt
@@ -1,15 +1,0 @@
-package com.example.android2023.data.response.weatherResponse
-
-
-import com.google.gson.annotations.SerializedName
-
-data class Weather(
-    @SerializedName("description")
-    val description: String,
-    @SerializedName("icon")
-    val icon: String,
-    @SerializedName("id")
-    val id: Int,
-    @SerializedName("main")
-    val main: String
-)
