--- app/src/main/java/com/example/android2023/data/response/weatherResponse/Wind.kt
@@ -1,13 +1,0 @@
-package com.example.android2023.data.response.weatherResponse
-
-
-import com.google.gson.annotations.SerializedName
-
-data class Wind(
-    @SerializedName("deg")
-    val deg: Int,
-    @SerializedName("gust")
-    val gust: Double,
-    @SerializedName("speed")
-    val speed: Double
-)
