+++ app/src/main/java/com/example/android2023/domain/usecase/GetWeatherByNameUseCase.kt
@@ -1,0 +1,12 @@
+package com.example.android2023.domain.usecase
+
+import com.example.android2023.data.datasource.remote.response.WeatherResponse
+import com.example.android2023.domain.WeatherRepository
+
+class GetWeatherByNameUseCase(
+    private val weatherRepository: WeatherRepository
+) {
+    suspend operator fun invoke(
+        query: String
+    ): WeatherResponse = weatherRepository.getWeatherByName(query)
+}
