+++ app/src/main/java/com/example/android2023/presentation/recyclerview/CityHolder.kt
@@ -1,4 +1,4 @@
-package com.example.android2023.ui.recyclerView
+package com.example.android2023.presentation.recyclerview
 
 import android.util.Log
 import android.view.LayoutInflater
@@ -6,7 +6,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import coil.load
 import com.example.android2023.databinding.ItemCityBinding
-import com.example.android2023.ui.recyclerView.models.CityItem
+import com.example.android2023.presentation.recyclerview.models.CityItem
 import timber.log.Timber
 
 class CityHolder(
@@ -21,7 +21,6 @@ class CityHolder(
 //            tvTemp.setTextColor(activity.getRecources.getColor(R.color.orange))
             tvTemp.setTextColor(binding.root.context.resources.getColor(city.tempColor()))
             ivIcon.load(city.url)
-
 
             //КАК СДЕЛАТЬ ВЕСЬ ITEM КЛИКАБЕЛЬНЫМ? работает через раз
             root.setOnClickListener {
